bootlinux: initial import of ansible role
This ansible role let's you get Linux from a git tag, compile, install
and reboot to use it. It is used on kdevops. This is an import with
the kdevops's bootlinux role as-is as of commit 459ae451097ece
("ansible: move devconfig role to ansible galaxy").

The only changes are to replace the symlinks on the configurations with
actual files. I suppose that does mean we will have to carry the config
twice if we want it elsewhere, but then again, why would you need the
config twice? I think we'll probably just get rid of the duplicates as
I do think we can live with the configs in this role alone. Even after
terraform support gets finalized with kdevops.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>