add configure linux failed_when clauses
Turn out we can return 0 or 141 on make oldconfig. Upon testing
we also cannot really tell when we really have changed, so just
use 0 or 141. Also, success can be 0 or 141.

The default for ansible is that any non zero return code means
a failure, so annotate when we expect it to succeed.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>