defaults: bump stable to v4.19.133
The v4.19.62 release has an issue which breaks compilation
with gcc 10 which changed the default to -fno-common, which
leads to this compilation failure:

    LD      arch/x86/boot/compressed/vmlinux
  ld: arch/x86/boot/compressed/pgtable_64.o:(.bss+0x0): multiple definition of `__force_order'; \
    arch/x86/boot/compressed/kaslr_64.o:(.bss+0x0): first defined here
  make[2]: *** [arch/x86/boot/compressed/Makefile:119: arch/x86/boot/compressed/vmlinux] Error 1

This is fixed by upstream commit df6d4f9db79c1a5d6f48b59db35ccd1e9ff9adfc
("x86/boot/compressed: Don't declare __force_order in kaslr_64.c") and
this got merged on the v4.19.y release as of the v4.19.109 release:

$ git describe --contains d71f8bd18cdfd8b5ea94acbd6248951d51e3770b
v4.19.109~60

But since 133 is newer just jump to that already.

This also add the respective configs for this kernel, and we
enable scsi-generic, and bridge to help with testing the block
layer, and for those who wish to work on nested virtualization.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>