tasks: add feature to infer target user and group based on login
Distributions / images can vary on the default username/group which,
and so if one is building a system to make this automatic, one either
has to add a whole bunch of conditional checks *per* image or
Linux distribution used *or*... you just infer the usernaem and
group based on the default login settings on target systems.

This enables the capacity to do the later, we infer your username
and group from *how* you log in to target systems. This allows
users of this role to be simplified considerably.

This feature is disabled by default to remain compatible with
older users of the bootlinux role.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>